+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.1-beta.1
+
+- Introduced new arguments in classes `AsyncElevatedButton`, `AsyncTextButton` and `AsyncOutlinedButton`:
+  - `loadingStyleBuilder`
+  - `successStyleBuilder`
+  - `failureStyleBuilder`
+- Animation smoothened for some `ButtonStyle`'s parameters
+
 ## 1.0.0-beta.1+1
 
 - Documentation corrections
@@ -8,6 +16,6 @@
 > Initial release
 
 - Introduced the following core classes:
-  - AsyncElevatedButton
-  - AsyncTextButton
-  - AsyncOutlinedButton
+  - `AsyncElevatedButton`
+  - `AsyncTextButton`
+  - `AsyncOutlinedButton`